fix(video): 修复 ui/pipeline + ui/other + 部分其他目录编译错误
ui/pipeline (121 errors):
- EpisodePlan Redeclaration → 重命名为 EpisodePlanRequest
- CostTrackerScreen: ifZero→if表达式, 类型统一, 返回类型注解
- NewProjectWizard: grid items/clip import, ProjectConfigPresets
- PipelineDashboardScreen: TextOverflow import, 返回类型注解
- PipelineStageScreen: infiniteRepeatableAnimation→infiniteRepeatable, progress lambda
- QualityReportScreen: 删除重复 displayName 扩展属性
- TemplateEngineScreen: grid imports, FontWeight.Small→Medium, LazyRow items嵌套修复

ui/other (205 errors):
- Icons import 修复 (settings/components/project/storyboard/preview/theme)
- VideoComponents: VideoTextField 组件修复
- VideoAnimations: animateFloat API 修复
- ProviderSettingsScreen: 大量 import 和类型修复
- StoryboardWorkbenchScreen: Icons 显式导入
- build.gradle: 添加 compose 依赖
- AppDatabase: SupportSQLiteDatabase import
- VideoProjectViewModel: appDb/类型修复
- 其他多个文件 import 和类型修复

+++ app/src/main/java/io/legado/app/video/ui/theme/VideoAnimations.kt
@@ -2,6 +2,7 @@ package io.legado.app.video.ui.theme
 
 import androidx.compose.animation.core.FastOutSlowInEasing
 import androidx.compose.animation.core.RepeatMode
+import androidx.compose.animation.core.animateFloat
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.infiniteRepeatable
 import androidx.compose.animation.core.rememberInfiniteTransition
@@ -130,7 +131,7 @@ fun WaveAnimation(
         Modifier.background(
             brush = Brush.verticalGradient(
                 colors = listOf(
-                    color.copy(alpha = 0.1f + 0.1f * kotlin.math.sin(waveOffset * Math.PI.toFloat())),
+                    color.copy(alpha = (0.1f + 0.1f * kotlin.math.sin(waveOffset * Math.PI.toFloat())).toFloat()),
                     color.copy(alpha = 0.3f)
                 )
             )
